change namespace for Handballnet Enums
