+++ contao/dca/tl_content.php
@@ -10,8 +10,8 @@ declare(strict_types=1);
  * @license MIT
  */
 
-use Janborg\H4aTabellen\HandballNet\Verband;
-use Janborg\H4aTabellen\HandballNet\Provider;
+use Janborg\H4aTabellen\HandballNet\Enum\Verband;
+use Janborg\H4aTabellen\HandballNet\Enum\Provider;
 use Janborg\H4aTabellen\Controller\ContentElement\HandballnetTabelleElement;
 use Janborg\H4aTabellen\Controller\ContentElement\HandballnetSpielplanElement;
 use Janborg\H4aTabellen\Controller\ContentElement\HandballnetWidgetElement;
