+++ src/HandballNet/DataTransferObject/TeamDto.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Janborg\H4aTabellen\HandballNet\DataTransferObject;
 
-use Janborg\H4aTabellen\HandballNet\AgeGroup;
-use Janborg\H4aTabellen\HandballNet\Provider;
-use Janborg\H4aTabellen\HandballNet\Verband;
+use Janborg\H4aTabellen\HandballNet\Enum\AgeGroup;
+use Janborg\H4aTabellen\HandballNet\Enum\Provider;
+use Janborg\H4aTabellen\HandballNet\Enum\Verband;
 
 final readonly class TeamDto
 {
