+++ src/HandballNet/Enum/AgeGroup.php
@@ -10,21 +10,28 @@ declare(strict_types=1);
  * @license MIT
  */
 
-namespace Janborg\H4aTabellen\HandballNet;
+namespace Janborg\H4aTabellen\HandballNet\Enum;
 
 use Contao\CoreBundle\Translation\TranslatableLabelInterface;
 use Symfony\Component\Translation\TranslatableMessage;
 
-enum Gender: string implements TranslatableLabelInterface
+enum AgeGroup: string implements TranslatableLabelInterface
 {
-    case MALE = 'Male';
-    case FEMALE = 'Female';
-    case MIXED = 'Mixed';
+    case MEN = 'Men';
+    case WOMEN = 'Women';
+    case AYOUTH = 'AYouth';
+    case BYOUTH = 'BYouth';
+    case CYOUTH = 'CYouth';
+    case DYOUTH = 'DYouth';
+    case EYOUTH = 'EYouth';
+    case FYOUTH = 'FYouth';
+    case MINIS = 'Minis';
+    case SENIORS = 'Seniors';
 
     public function label(): TranslatableMessage
     {
         return new TranslatableMessage(
-            'gender.label.'.$this->value,
+            'agegroup.label.'.$this->value,
             [],
             'handballnet',
         );
