+++ src/HandballNet/Enum/Gender.php
@@ -10,28 +10,21 @@ declare(strict_types=1);
  * @license MIT
  */
 
-namespace Janborg\H4aTabellen\HandballNet;
+namespace Janborg\H4aTabellen\HandballNet\Enum;
 
 use Contao\CoreBundle\Translation\TranslatableLabelInterface;
 use Symfony\Component\Translation\TranslatableMessage;
 
-enum AgeGroup: string implements TranslatableLabelInterface
+enum Gender: string implements TranslatableLabelInterface
 {
-    case MEN = 'Men';
-    case WOMEN = 'Women';
-    case AYOUTH = 'AYouth';
-    case BYOUTH = 'BYouth';
-    case CYOUTH = 'CYouth';
-    case DYOUTH = 'DYouth';
-    case EYOUTH = 'EYouth';
-    case FYOUTH = 'FYouth';
-    case MINIS = 'Minis';
-    case SENIORS = 'Seniors';
+    case MALE = 'Male';
+    case FEMALE = 'Female';
+    case MIXED = 'Mixed';
 
     public function label(): TranslatableMessage
     {
         return new TranslatableMessage(
-            'agegroup.label.'.$this->value,
+            'gender.label.'.$this->value,
             [],
             'handballnet',
         );
