+++ src/Model/HandballnetTeamsModel.php
@@ -14,9 +14,9 @@ namespace Janborg\H4aTabellen\Model;
 
 use Contao\Model;
 use Contao\Model\Collection;
-use Janborg\H4aTabellen\HandballNet\AgeGroup;
-use Janborg\H4aTabellen\HandballNet\Provider;
-use Janborg\H4aTabellen\HandballNet\Verband;
+use Janborg\H4aTabellen\HandballNet\Enum\AgeGroup;
+use Janborg\H4aTabellen\HandballNet\Enum\Provider;
+use Janborg\H4aTabellen\HandballNet\Enum\Verband;
 
 /**
  * Reads and writes handballnetTeams.
